refactor: change file paths, delete files
