+++ src/shared/types/pokemon.types.ts
